- Fix for calculation of final WaveTable value due to float precision loss through erroneous int32 conversion & potential missed FinalValue copy on resave #rnx #tests local using sample MetaSound and printlog #rb helen.yang
[CL 29604416 by rob gay in ue5-main branch]

+++ Engine/Plugins/Runtime/WaveTable/Source/WaveTable/Private/WaveTableSampler.cpp
@@ -480,7 +480,7 @@ namespace WaveTable
 		}
 	}
 
-	int32 FWaveTableSampler::ComputeIndexFinished(TArrayView<const float> InSyncTriggers, TArrayView<float> OutIndicesView)
+	float FWaveTableSampler::ComputeIndexFinished(TArrayView<const float> InSyncTriggers, TArrayView<float> OutIndicesView)
 	{
 		check(!OutIndicesView.IsEmpty());
